провёл refactoring кода, переписал тесты
