+++ src/test/java/school/faang/user_service/service/SubscriptionServiceTest.java
@@ -1,9 +1,11 @@
 package school.faang.user_service.service;
 
 import org.jetbrains.annotations.NotNull;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.repository.SubscriptionRepository;
 
@@ -19,31 +21,49 @@ public class SubscriptionServiceTest {
     @Mock
     private SubscriptionRepository repository;
 
+    @BeforeEach
+    void setUp() {
+        MockitoAnnotations.openMocks(this);
+    }
+
     @Test
-    void testRepeatSubscription() {
+    void testSubscriptionForYourself() {
+        long followerId = 1L;
+        long followeeId = 1L;
+        assertThrows(DataValidationException.class, () -> service.followUser(followerId, followeeId));
+    }
+
+    @Test
+    void testRepeatSubscriptionForAnotherUser() {
         InitialData data = getData(true);
         assertThrows(DataValidationException.class, () -> service.followUser(data.followerId(), data.followeeId()));
     }
 
     @Test
-    void testSubscription() {
+    void testSubscriptionForAnotherUser() {
         InitialData data = getData(false);
         service.followUser(data.followerId(), data.followeeId());
         verify(repository, times(1)).followUser(data.followerId(), data.followeeId());
     }
 
     @Test
-    void testUnsubscription() {
+    void testUnsubscribingFromYourself() {
+        long followerId = 1L;
+        long followeeId = 1L;
+        assertThrows(DataValidationException.class, () -> service.unfollowUser(followerId, followeeId));
+    }
+
+    @Test
+    void testNonExistingUnsubscriptionFromAnotherUser() {
+        InitialData data = getData(false);
+        assertThrows(DataValidationException.class, () -> service.unfollowUser(data.followerId(), data.followeeId()));
+    }
+
+    @Test
+    void testUnsubscriptionFromAnotherUser() {
         InitialData data = getData(true);
         service.unfollowUser(data.followerId(), data.followeeId());
         verify(repository, times(1)).unfollowUser(data.followerId(), data.followeeId());
-    }
-
-    // протестировать повторную отписку
-    @Test
-    void testRepeatUnsubscription() {
-        InitialData data = getData(false);
-        assertThrows(DataValidationException.class, () -> service.unfollowUser(data.followerId(), data.followeeId()));
     }
 
     private @NotNull InitialData getData(boolean isThereSub) {
